=:完善登录

+++ src/main/java/com/yochalyc/myblog/blog/web/model/Token.java
@@ -22,6 +22,6 @@ public class Token {
     /**
      * expire interval
      */
-    private long exp;
+    private Integer exp;
 
 }
